+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Version 8.0.0 (2021-10-12)
+## Version 8.0.0 (2021-11-16)
 
 ### Features
 
@@ -48,6 +48,8 @@
 - added new `isNotType` function that gets indication whether is provided `Type` of `NotType`
 - added new `ParametersTransformsObj` interface that defines object for parameter transforms
 - added new decorator `PATCH` allowin to create *PATCH* http method builder
+- `ResponseTransform` decorator now also takes function or array of functions
+- `ParameterTransform` decorator now also takes function or array of functions
 - *response transform* now also gets input *arguments*
 - added new types that are used for helping working with metadata stored in *Descriptor*
     - `RestHttpHeaders` - contains additional headers that will be added